explaining in a paragraph why it is useful to verify distributed algorithms ..as they are designed and proved by hand. Then, in a second paragraph a short explanation about the difference of automatic and deductive verification, and how it is unfeasible to apply automatic verification to distributed algorithms (due to them being parameterized, and parameterized verification is undecidable in general [1]).
Finally, say how Ivy is tailored to dealing with
these kinds of algorithms, and is a tool that allows the user to provide
invariants and check them automatically.

Also, an outline of the rest of the
thesis, explaining in a few words what each section is about can be added at
the end of the introduction.